+++ redismanager-dao/src/main/java/com/tz/redismanager/dao/domain/po/UserRoleRelationPO.java
@@ -21,17 +21,17 @@ public class UserRoleRelationPO {
     /**
      * 主键
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 用户Id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 角色id
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 创建人
